executor: add process executor
* split src.ewoksorange.gui.concurrency.EwoksExecutor to a 'executor' module
* add:

  * ProcessCallable: Top-level picklable callable submitted to ProcessPoolExecutor.
  * EwoksProcessWorker: worker for the process
