+++ src/ewoksorange/gui/concurrency/executor/_EwoksThreadWorker.py
@@ -1,0 +1,41 @@
+from threading import Lock
+from typing import Optional
+from typing import Type
+
+from ewokscore import TaskWithProgress
+from ewokscore.task import Task
+
+
+class EwoksThreadWorker:
+    """Callable that instantiates and executes an ewoks task in the worker thread."""
+
+    def __init__(self, task_class: Type[Task], **task_kwargs):
+        self._task_class = task_class
+        self._task_kwargs = task_kwargs
+        self._task: Optional[Task] = None
+        self._lock = Lock()
+
+    def __call__(self):
+        task_class = self._task_class
+        kwargs = dict(self._task_kwargs)
+        if not issubclass(task_class, TaskWithProgress):
+            kwargs.pop("progress", None)
+
+        task = task_class(**kwargs)
+        with self._lock:
+            self._task = task
+
+        task.execute()
+        return task.output_variables
+
+    def abort(self) -> None:
+        """Call the ewoks task's cancel() to stop a running task."""
+        with self._lock:
+            task = self._task
+            if task is not None:
+                task.cancel()
+
+    @property
+    def has_task(self) -> bool:
+        with self._lock:
+            return self._task is not None
