+++ src/ewoksorange/gui/concurrency/executor/__init__.py
@@ -1,0 +1,2 @@
+from .EwoksExecutor import EwoksExecutor  # noqa F401
+from .TaskFuture import TaskFuture  # noqa F401
